feat: lotto번호 일치 확인 기능 추가
